Refactor all tests
