+++ test/E2ETest/Microsoft.Test.E2E.AspNet.OData/Build.AspNetCore3x/Microsoft.Test.E2E.AspNetCore3x.OData.csproj
@@ -1665,18 +1665,21 @@
     <Compile Include="..\Validation\DeltaOfTValidationTests.cs">
       <Link>Validation\DeltaOfTValidationTests.cs</Link>
     </Compile>
-	<Compile Include="..\BulkOperation\BulkInsertDataModel.cs">
-		<Link>BulkOperation\BulkInsertDataModel.cs</Link>
-	</Compile>
-	<Compile Include="..\BulkOperation\BulkInsertEdmModel.cs">
-		<Link>BulkOperation\BulkInsertEdmModel.cs</Link>
-	</Compile>
-	<Compile Include="..\BulkOperation\BulkInsertController.cs">
-		<Link>BulkOperation\BulkInsertController.cs</Link>
-	</Compile>
-	<Compile Include="..\BulkOperation\BulkOperationPatchHandlers.cs">
-		<Link>BulkOperation\BulkOperationPatchHandlers.cs</Link>
-	</Compile>
+    <Compile Include="..\BulkOperation\BulkInsertTest.cs">
+      <Link>BulkOperation\BulkInsertTest.cs</Link>
+    </Compile>
+    <Compile Include="..\BulkOperation\BulkInsertDataModel.cs">
+      <Link>BulkOperation\BulkInsertDataModel.cs</Link>
+    </Compile>
+    <Compile Include="..\BulkOperation\BulkInsertEdmModel.cs">
+      <Link>BulkOperation\BulkInsertEdmModel.cs</Link>
+    </Compile>
+    <Compile Include="..\BulkOperation\BulkInsertController.cs">
+      <Link>BulkOperation\BulkInsertController.cs</Link>
+    </Compile>
+    <Compile Include="..\BulkOperation\BulkOperationPatchHandlers.cs">
+      <Link>BulkOperation\BulkOperationPatchHandlers.cs</Link>
+    </Compile>
   </ItemGroup>
   
   <ItemGroup>
